+++ purchase_backorder/__manifest__.py
@@ -4,7 +4,7 @@
 
 {
     "name": "Purchase Backorder Report",
-    "version": "15.0.1.0.0",
+    "version": "16.0.1.0.0",
     "license": "AGPL-3",
     "author": "Open Source Integrators, Odoo Community Association (OCA)",
     "summary": "Report of Un-Invoiced Goods Received and Backorders",
