+++ purchase_backorder/models/__init__.py
@@ -2,4 +2,5 @@
 # Copyright (C) 2021 Serpent Consulting Services
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import purchase
+from . import purchase_order
+from . import purchase_order_line
