+++ purchase_backorder/readme/CONTRIBUTORS.rst
@@ -2,3 +2,4 @@
 * Sandip Mangukiya <smangukiya@opensourceintegrators.com>
 * Bhavesh Odedra <bodedra@opensourceintegrators.com>
 * Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+* Murtaza Mithaiwala <mmithaiwala@opensourceintegrators.com>
